Added Admin approval page where admin can manage staff state for users and can search them by nicknames and delete them

+++ parser/views.py
@@ -1,8 +1,9 @@
-from django.shortcuts import render
+from django.shortcuts import render, redirect
 from django.core.paginator import Paginator
 from django.http import HttpResponseRedirect, HttpResponse
 from django.db.models import Q
 from django.urls import reverse
+from django.contrib import messages
 from .parser import parse, kill_parse, SITE_MESSAGES
 from .models import Forum, Nickname, BannedFilter, KeyWordFilter
 import json
@@ -16,10 +17,14 @@ def index(request):
     return render(request, 'parser/index.html')
 
 def chooseFromSaved(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     return render(request, 'parser/choose_from_saved.html')
 
 def tagDataGet(request):
-
+    if not request.user.is_authenticated:
+        return redirect('about')
+        
     if request.method == 'POST':
         regex_pattern = '/~/'
         message_parameter = {}
@@ -85,6 +90,8 @@ def about(request):
     return render(request, 'parser/about.html')
 
 def resultPage(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     p = Paginator(Nickname.objects.all(), 100)
     page = request.GET.get('page')
     users = p.get_page(page)
@@ -98,6 +105,8 @@ def resultPage(request):
     return render(request, 'parser/results.html', context)
 
 def parsing(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     context = {
         'forums': Forum.objects.all().order_by('link'),
         'filters' : BannedFilter.objects.all().order_by('purpose'),
@@ -110,6 +119,8 @@ def parsing(request):
     return render(request, 'parser/parse_page.html', context)
 
 def save_result(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     response = HttpResponse(content_type='text/csv')
     response['Content-Disposition'] = 'attachment; filename=result_user.csv'
     writer = csv.writer(response)
@@ -120,6 +131,8 @@ def save_result(request):
     return response
 
 def save_search_results(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     response = HttpResponse(content_type='text/csv')
     response['Content-Disposition'] = 'attachment; filename=result_user.csv'
     writer = csv.writer(response)
@@ -134,7 +147,10 @@ def save_search_results(request):
             writer.writerow([user.id, user.handler, user.user_id, user.nicknames, user.forumOrigin,])
     return response
 
+
 def save_as_excel(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     forums=(set(Forum.objects.values_list('link')))
     forumlist={}
     response = HttpResponse(content_type='text/xslx')
@@ -145,16 +161,13 @@ def save_as_excel(request):
         for forum in forums:
             if str(user.forumOrigin) in str(forum):
                 name = user.forumOrigin.link.replace('/', '')
-                print(name)
                 name = name.replace('http:', '')
                 name = name.replace('https:', '')
-                print(name)
                 try:
                     forumlist[f'{name}'].append([user.handler, user.nicknames])
                 except:
                     newadd={f'{name}':[[user.handler, user.nicknames]]}
                     forumlist.update(newadd)
-    
     for list in forumlist:
         DataFrame(forumlist[list]).to_excel(writer,sheet_name=f'{list}',index=False)
     writer.save()                   
@@ -170,6 +183,8 @@ def save_as_excel(request):
 
 # ============================================[ HTMX functions ]============================================ #
 def tagDataNewField(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     if request.method == 'GET':
         parameter = request.GET.get('parameter')
         value = int(request.GET.get('value'))
@@ -187,6 +202,8 @@ def tagDataNewField(request):
         return render(request, 'parser/partials/formField.html', context)
 
 def startParse(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     context = {}
     if request.method == 'POST':
         link = request.POST.get('linkSelect')
@@ -199,6 +216,8 @@ def startParse(request):
         return render(request, 'parser/partials/stopButton.html', context)
 
 def stopParse(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     if request.method == "PATCH":
         kill_parse()
         context = {
@@ -208,6 +227,8 @@ def stopParse(request):
         return render(request, 'parser/partials/parsingForm.html', context)
     
 def getProgressText(request):
+    if not request.user.is_authenticated:
+        return redirect('about')
     text_to_pass = ''
     if request.method == "GET":
         if not len(SITE_MESSAGES) == 0:
@@ -217,7 +238,9 @@ def getProgressText(request):
         else:
             return HttpResponse("")
 
-def search_user(request):      
+def search_user(request):
+    if not request.user.is_authenticated:
+        return redirect('about')      
     search_text = request.POST.get('search')
     forum_of_origin = request.POST.get('forumfilter')
     if search_text == "" and forum_of_origin == "":
@@ -234,6 +257,8 @@ def search_user(request):
         return render(request, 'parser/partials/search_results.html', context)
 
 def delete_handler(request, id):
+    if not request.user.is_authenticated:
+        return redirect('about')
     if request.method == "GET":
         Nickname.objects.filter(pk=id).delete()
         response = HttpResponse("Okay")
@@ -241,6 +266,8 @@ def delete_handler(request, id):
         return response
 
 def add_to_keys(request, id):
+    if not request.user.is_authenticated:
+        return redirect('about')
     if request.method == "GET":
         key_word_db = KeyWordFilter.objects.get(purpose='Telegram')
         key_word_db.filter += '\n' + Nickname.objects.get(pk=id).handler
@@ -251,6 +278,8 @@ def add_to_keys(request, id):
         return response
 
 def add_to_banned(request, id):
+    if not request.user.is_authenticated:
+        return redirect('about')
     if request.method == "GET":
         banned_word_db = BannedFilter.objects.get(purpose='Telegram')
         banned_word_db.filter += '\n' + Nickname.objects.get(pk=id).handler
